+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export * from './layout'
 export * from './popups'
-export * from './ui/custom'
+export * from './ui'
