--- src/components/ui/custom/Card/index.tsx
@@ -1,86 +1,0 @@
-import * as React from "react";
-
-import { cn } from "@/config";
-
-const Card = React.forwardRef<
-  HTMLDivElement,
-  React.HTMLAttributes<HTMLDivElement>
->( ( { className, ...props }, ref ) => ( 
-    <div
-        ref={ref}
-        className={cn(
-            "rounded-lg border border-slate-200 bg-white text-slate-950 shadow-sm dark:border-slate-800 dark:bg-slate-950 dark:text-slate-50",
-            className,
-        )}
-        {...props}
-    />
-) );
-Card.displayName = "Card";
-
-const CardHeader = React.forwardRef<
-  HTMLDivElement,
-  React.HTMLAttributes<HTMLDivElement>
->( ( { className, ...props }, ref ) => (
-    <div
-        ref={ref}
-        className={cn( "flex flex-col space-y-1.5 p-6", className )}
-        {...props}
-    />
-) );
-CardHeader.displayName = "CardHeader";
-
-const CardTitle = React.forwardRef<
-  HTMLParagraphElement,
-  React.HTMLAttributes<HTMLHeadingElement>
->( ( { className, ...props }, ref ) => (
-    <h3
-        ref={ref}
-        className={cn(
-            "text-2xl font-semibold leading-none tracking-tight",
-            className,
-        )}
-        {...props}
-    />
-) );
-CardTitle.displayName = "CardTitle";
-
-const CardDescription = React.forwardRef<
-  HTMLParagraphElement,
-  React.HTMLAttributes<HTMLParagraphElement>
->( ( { className, ...props }, ref ) => (
-    <p
-        ref={ref}
-        className={cn( "text-sm text-neutral-900 dark:text-neutral-900", className )}
-        {...props}
-    />
-) );
-CardDescription.displayName = "CardDescription";
-
-const CardContent = React.forwardRef<
-  HTMLDivElement,
-  React.HTMLAttributes<HTMLDivElement>
->( ( { className, ...props }, ref ) => (
-    <div ref={ref} className={cn( "p-6 pt-0", className )} {...props} />
-) );
-CardContent.displayName = "CardContent";
-
-const CardFooter = React.forwardRef<
-  HTMLDivElement,
-  React.HTMLAttributes<HTMLDivElement>
->( ( { className, ...props }, ref ) => (
-    <div
-        ref={ref}
-        className={cn( "flex items-center p-6 pt-0", className )}
-        {...props}
-    />
-) );
-CardFooter.displayName = "CardFooter";
-
-export {
-    Card,
-    CardContent,
-    CardDescription,
-    CardFooter,
-    CardHeader,
-    CardTitle,
-};
