--- src/components/ui/custom/Input/index.tsx
@@ -1,25 +1,0 @@
-import * as React from "react";
-
-import { cn } from "@/config";
-
-export interface InputProps
-  extends React.InputHTMLAttributes<HTMLInputElement> {}
-
-const Input = React.forwardRef<HTMLInputElement, InputProps>(
-    ( { className, type, ...props }, ref ) => {
-        return (
-            <input
-                type={type}
-                className={cn(
-                    "flex h-10 w-full rounded border border-neutral-500 bg-white px-3 py-2 ring-offset-white file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-slate-500 focus-visible:outline-none focus-visible:ring-slate-950 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 dark:border-slate-800 dark:bg-slate-950  dark:ring-offset-slate-950 dark:placeholder:text-slate-400 dark:focus-visible:ring-slate-300",
-                    className,
-                )}
-                ref={ref}
-                {...props}
-            />
-        );
-    },
-);
-Input.displayName = "Input";
-
-export { Input };
