--- src/components/ui/custom/index.ts
@@ -1,19 +1,0 @@
-export * from "./Body/index";
-export * from "./Button/index";
-export * from "./CancelIcon/index";
-export * from "./Card/index";
-export * from "./CodeVerification/index";
-export * from "./Container/index";
-export * from "./Form/index";
-export * from "./FormError/index";
-export * from "./FormLoading/index";
-export * from "./FormSuccess/index";
-export * from "./Input/index";
-export * from "./Label/index";
-export * from "./Line/index";
-export * from "./LoadingInputField/index";
-export * from "./LoadingLabel/index";
-export * from "./LoadingPulse/index";
-export * from "./Logo/index";
-export * from "./LogoutButton/index";
-export * from "./SubLineHeading/index";
